+++ R/app_server.R
@@ -99,8 +99,6 @@ app_server <- function(
     }
   })
   
-  forms_to_review_data <- app_vars$form_level_data[c("item_group", "review_required")] 
-  
   observeEvent(user_error(), {
     showNotification(
       user_error(), 
@@ -130,21 +128,14 @@ app_server <- function(
   )
   
   start_page_summary_vars <- c("subject_status", "WHO.classification", "Age", "Sex", "event_name")
+  forms_to_review <- with(app_vars$form_level_data, item_group[review_required])
   rev_data <- reactiveValues(
     summary = reactive({
-      req(forms_to_review_data)
-      r$review_data |>
-        reactiveValuesToList() |> 
-        do.call(what = rbind) |> 
-        dplyr::left_join(forms_to_review_data, by = "item_group") |> 
-        dplyr::filter(
-          reviewed != "Yes",
-          review_required,
-          subject_id %in% r$filtered_subjects
-        ) |>
-        summarize_review_data() |>
-        dplyr::select(subject_id, "Form" = item_group, "Event" = event_name,
-                      "Edit date" = edit_date_time, status, reviewed)
+      req(forms_to_review)
+      reactiveValuesToList(r$review_data)[forms_to_review] |> 
+        dplyr::bind_rows() |> 
+        subset(reviewed != "Yes" & subject_id %in% r$filtered_subjects) |> 
+        summarize_review_data()
     }),
     overview = reactive({
       with(static_overview_data, static_overview_data[subject_id %in% r$filtered_subjects, ]) |>
@@ -349,7 +340,7 @@ app_server <- function(
       app_vars = app_vars,
       navinfo,
       forms_to_review = reactive({
-        with(rev_data$summary(), Form[subject_id == r$subject_id])
+        with(rev_data$summary(), item_group[subject_id == r$subject_id])
       }),
       db_path = user_db,
       available_data = available_data