+++ tests/testthat/test-mod_header_widgets.R
@@ -82,7 +82,7 @@ describe(
             summary = reactive({
               data.frame(
                 "subject_id" = "Subj01",
-                "Form" = "Adverse events",
+                "item_group" = "Adverse events",
                 reviewed = c("No", "Yes")
               )
             })
@@ -154,7 +154,7 @@ describe(
             summary = reactive({
               data.frame(
                 "subject_id" = "Subj01",
-                "Form" = "Adverse events",
+                "item_group" = "Adverse events",
                 reviewed = c("No", "Yes")
               )
             })