+++ NEWS.md
@@ -23,6 +23,7 @@
 
 - Updated role of long-term contributors to co-authors in Description field (#246).
 - Refactored some functions so that it is no longer needed to create an `apptables` object when starting the applcation, improving start up efficiency (#251).
+- Refactored function to retrieve summary data to improve efficiency.
 
 # clinsight 0.3.0